Refactor Address to rentalinformation package
